Update buildexpression code with new implementation
