+++ pkg/platform/api/buildplanner/model/buildexpression.go
@@ -3,10 +3,13 @@ package model
 import (
 	"encoding/json"
 	"errors"
+	"fmt"
+	"sort"
+	"strings"
 	"time"
 
 	"github.com/ActiveState/cli/internal/errs"
-	"github.com/ActiveState/cli/internal/logging"
+	"github.com/ActiveState/cli/internal/rtutils/p"
 	"github.com/go-openapi/strfmt"
 )
 
@@ -24,12 +27,22 @@ const (
 	OperationRemoved
 	OperationUpdated
 
-	SolveFuncName           = "solve"
-	SolveLegacyFuncName     = "solve_legacy"
-	RequirementsKey         = "requirements"
-	AtTimeKey               = "at_time"
-	RequirementNameKey      = "name"
-	RequirementNamespaceKey = "namespace"
+	SolveFuncName                     = "solve"
+	SolveLegacyFuncName               = "solve_legacy"
+	RequirementsKey                   = "requirements"
+	AtTimeKey                         = "at_time"
+	RequirementNameKey                = "name"
+	RequirementNamespaceKey           = "namespace"
+	RequirementVersionRequirementsKey = "version_requirements"
+	RequirementVersionKey             = "version"
+	RequirementComparatorKey          = "comparator"
+
+	ctxLet         = "let"
+	ctxIn          = "in"
+	ctxAp          = "ap"
+	ctxValue       = "value"
+	ctxAssignments = "assignments"
+	ctxIsAp        = "isAp"
 )
 
 func (o Operation) String() string {
@@ -47,192 +60,6 @@ func (o Operation) String() string {
 
 var funcNodeNotFoundError = errors.New("Could not find function node")
 
-type BuildExpression2 struct {
-	Lets []Let
-	Aps  []Ap
-	Vars []Var
-	AST  map[string]interface{}
-}
-
-type Ap struct {
-	Name      string
-	Arguments map[string]interface{}
-}
-
-type Let struct {
-	Arguments map[string]interface{}
-	InExpr    interface{}
-}
-
-type Var struct {
-	Name  string
-	Value interface{}
-}
-
-func NewBuildExpression2(data []byte) (*BuildExpression2, error) {
-	expressionMap := make(map[string]interface{})
-	err := json.Unmarshal(data, &expressionMap)
-	if err != nil {
-		return nil, errs.Wrap(err, "Could not unmarshal JSON")
-	}
-
-	// TODO: See if we can use a modified version of Mitchell's code to build the AST
-	// If we can then discuss with the team how to integrate these changes together or
-	// file a follow up story
-	resultExpression := BuildExpression2{}
-	err = traverse(expressionMap, &resultExpression)
-	if err != nil {
-		return nil, errs.Wrap(err, "Could not build expression")
-	}
-
-	return &resultExpression, nil
-}
-
-// func traverseForAst(expressionMap, ast map[string]interface{}) error {
-// 	for key, value := range expressionMap {
-// 		switch {
-// 		case isLet(key, value):
-// 			rawArguments, ok := value.(map[string]interface{})
-// 			if !ok {
-// 				return errs.New("Let value is not a map")
-// 			}
-
-// 			// Build up the AST from the let value
-// 			arguments := make(map[string]interface{})
-// 			err := traverseForAst(rawArguments, arguments)
-// 			if err != nil {
-// 				return errs.Wrap(err, "Could not build expression")
-// 			}
-
-// 			var inExpr interface{}
-// 			rawInExpr, ok := rawArguments["in"].(map[string]interface{})
-// 			if ok {
-// 				inExpr = make(map[string]interface{})
-// 				err = traverseForAst(rawInExpr, inExpr)
-// 				if err != nil {
-// 					return errs.Wrap(err, "Could not build expression")
-// 				}
-// 			}
-
-// 			let := Let{
-// 				Arguments: value.(map[string]interface{}),
-// 				InExpr:    rawArguments["in"],
-// 			}
-// 		}
-// 	}
-
-// 	return nil
-// }
-
-func traverse(expressionMap map[string]interface{}, result *BuildExpression2) error {
-	for key, value := range expressionMap {
-		_, valueIsMap := value.(map[string]interface{})
-		switch {
-		case isLet(key, value):
-			letVal, ok := value.(map[string]interface{})
-			if !ok {
-				return errs.New("Let value is not a map")
-			}
-
-			let := Let{
-				Arguments: value.(map[string]interface{}),
-				InExpr:    letVal["in"],
-			}
-
-			result.Lets = append(result.Lets, let)
-			result.AST[key] = let
-
-			err := traverse(value.(map[string]interface{}), result)
-			if err != nil {
-				return errs.Wrap(err, "Could not build expression")
-			}
-		case isAp(value):
-			var apName string
-			for k := range value.(map[string]interface{}) {
-				apName = k
-			}
-
-			args, ok := value.(map[string]interface{})[apName].(map[string]interface{})
-			if !ok {
-				return errs.New("Ap arguments are not a map")
-			}
-
-			ap := Ap{
-				Name:      apName,
-				Arguments: args,
-			}
-
-			result.Aps = append(result.Aps, ap)
-		case isVar(value):
-			// May need some sort of context so we can identify the runtime variable with it's value
-			// being the solve_legacy ap
-			variable := Var{
-				Name:  key,
-				Value: value,
-			}
-
-			result.Vars = append(result.Vars, variable)
-		case valueIsMap:
-			err := traverse(value.(map[string]interface{}), result)
-			if err != nil {
-				return errs.Wrap(err, "Could not build expression")
-			}
-		}
-	}
-
-	return nil
-}
-
-func isLet(key string, value interface{}) bool {
-	if key == "let" {
-		letMap, ok := value.(map[string]interface{})
-		if !ok {
-			return false
-		}
-		if _, ok := letMap["in"]; !ok {
-			return false
-		}
-
-		return true
-	}
-	return false
-}
-
-func isAp(value interface{}) bool {
-	apMap, ok := value.(map[string]interface{})
-	if !ok {
-		return false
-	}
-
-	if len(apMap) != 1 {
-		return false
-	}
-
-	var apName string
-	for k := range apMap {
-		apName = k
-	}
-
-	argMap, ok := apMap[apName].(map[string]interface{})
-	if !ok {
-		return false
-	}
-
-	if _, ok := argMap["in"]; ok {
-		return false
-	}
-
-	return true
-}
-
-func isVar(value interface{}) bool {
-	if _, ok := value.(string); ok {
-		return true
-	}
-
-	return false
-}
-
 type Requirement struct {
 	Name               string               `json:"name"`
 	Namespace          string               `json:"namespace"`
@@ -242,9 +69,66 @@ type Requirement struct {
 type VersionRequirement map[string]string
 
 type BuildExpression struct {
-	expression   map[string]interface{}
-	solveNode    *map[string]interface{}
-	requirements []Requirement
+	Let *Let
+}
+
+type Let struct {
+	Assignments []*Var
+	In          *In
+}
+
+type Var struct {
+	Name  string
+	Value *Value
+}
+
+type Value struct {
+	Ap   *Ap
+	List *[]*Value
+	Str  *string
+	Null *Null
+
+	Assignment *Var
+	Object     *[]*Var
+	Ident      *string
+}
+
+type Null struct {
+	Null string
+}
+
+type Ap struct {
+	Name      string
+	Arguments []*Value
+}
+
+type In struct {
+	FuncCall *Ap
+	Name     *string
+}
+
+type context []string
+
+func (s *context) push(str string) {
+	*s = append(*s, str)
+}
+
+func (s *context) pop() (string, error) {
+	if len(*s) == 0 {
+		return "", errs.New("stack is empty")
+	}
+	str := (*s)[len(*s)-1]
+	*s = (*s)[:len(*s)-1]
+	return str, nil
+}
+
+func (s *context) contains(str string) bool {
+	for _, v := range *s {
+		if v == str {
+			return true
+		}
+	}
+	return false
 }
 
 // NewBuildExpression creates a BuildExpression from a JSON byte array.
@@ -283,122 +167,246 @@ type BuildExpression struct {
 //	  }
 //	}
 func NewBuildExpression(data []byte) (*BuildExpression, error) {
-	expression := make(map[string]interface{})
-	err := json.Unmarshal(data, &expression)
+	m := make(map[string]interface{})
+	err := json.Unmarshal(data, &m)
 	if err != nil {
-		return nil, errs.Wrap(err, "Could not unmarshal JSON")
+		return nil, errs.Wrap(err, "Could not unmarshal build expression")
 	}
 
-	solveNode, err := getSolveNode(expression)
-	if err != nil {
-		return nil, errs.Wrap(err, "Could not get solve node")
+	letValue, ok := m["let"]
+	if !ok {
+		return nil, errs.New("Build expression has no 'let' key")
+	}
+	letMap, ok := letValue.(map[string]interface{})
+	if !ok {
+		return nil, errs.New("'let' key is not a JSON object")
 	}
 
-	requirements, err := getRequirements(solveNode)
+	ctx := context{}
+	let, err := newLet(ctx, letMap)
 	if err != nil {
-		return nil, errs.Wrap(err, "Could not get requirements node")
+		return nil, errs.Wrap(err, "Could not parse 'let' key")
 	}
 
-	return &BuildExpression{
-		expression:   expression,
-		solveNode:    &solveNode,
-		requirements: requirements,
-	}, nil
+	expr := &BuildExpression{let}
+
+	err = expr.validateRequirements()
+	if err != nil {
+		return nil, errs.Wrap(err, "Could not validate requirements")
+	}
+
+	return expr, nil
+}
+
+func newLet(ctx context, m map[string]interface{}) (*Let, error) {
+	ctx.push(ctxLet)
+	defer ctx.pop()
+
+	inValue, ok := m["in"]
+	if !ok {
+		return nil, errs.New("Build expression's 'let' object has no 'in' key")
+	}
+
+	in, err := newIn(ctx, inValue)
+	if err != nil {
+		return nil, errs.Wrap(err, "Could not parse 'in' key's value: %v", inValue)
+	}
+
+	// Delete in so it doesn't get parsed as an assignment.
+	delete(m, "in")
+
+	assignments, err := newAssignments(ctx, m)
+	if err != nil {
+		return nil, errs.Wrap(err, "Could not parse 'let' key")
+	}
+
+	return &Let{Assignments: *assignments, In: in}, nil
+}
+
+func isAp(ctx context, value map[string]interface{}) bool {
+	ctx.push(ctxIsAp)
+	defer ctx.pop()
+
+	_, hasIn := value["in"]
+	if hasIn && !ctx.contains(ctxAssignments) {
+		return false
+	}
+
+	return true
+}
+
+func newValue(ctx context, valueInterface interface{}) (*Value, error) {
+	ctx.push(ctxValue)
+	defer ctx.pop()
+
+	value := &Value{}
+
+	switch v := valueInterface.(type) {
+	case map[string]interface{}:
+		// Examine keys first to see if this is a function call.
+		for key, val := range v {
+			if _, ok := val.(map[string]interface{}); !ok {
+				continue
+			}
+
+			if isAp(ctx, val.(map[string]interface{})) {
+				f, err := newAp(ctx, v)
+				if err != nil {
+					return nil, errs.Wrap(err, "Could not parse '%s' function's value: %v", key, v)
+				}
+				value.Ap = f
+			}
+		}
+
+		if value.Ap == nil {
+			// It's not a function call, but an object.
+			object, err := newAssignments(ctx, v)
+			if err != nil {
+				return nil, errs.Wrap(err, "Could not parse object: %v", v)
+			}
+			value.Object = object
+		}
+
+	case []interface{}:
+		values := []*Value{}
+		for _, item := range v {
+			value, err := newValue(ctx, item)
+			if err != nil {
+				return nil, errs.Wrap(err, "Could not parse list: %v", v)
+			}
+			values = append(values, value)
+		}
+		value.List = &values
+
+	case string:
+		if ctx.contains(ctxIn) {
+			value.Ident = &v
+		} else {
+			value.Str = p.StrP(v)
+		}
+
+	default:
+		// An empty value is interpreted as JSON null.
+		value.Null = &Null{}
+	}
+
+	return value, nil
+}
+
+func newAp(ctx context, m map[string]interface{}) (*Ap, error) {
+	ctx.push(ctxAp)
+	defer ctx.pop()
+
+	// Look in the given object for the function's name and argument object or list.
+	var name string
+	var argsInterface interface{}
+	for key, value := range m {
+		if isAp(ctx, value.(map[string]interface{})) {
+			name = key
+			argsInterface = value
+			break
+		}
+	}
+
+	args := []*Value{}
+
+	switch v := argsInterface.(type) {
+	case map[string]interface{}:
+		for key, valueInterface := range v {
+			value, err := newValue(ctx, valueInterface)
+			if err != nil {
+				return nil, errs.Wrap(err, "Could not parse '%s' function's argument '%s': %v", name, key, valueInterface)
+			}
+			args = append(args, &Value{Assignment: &Var{Name: key, Value: value}})
+		}
+		sort.SliceStable(args, func(i, j int) bool { return args[i].Assignment.Name < args[j].Assignment.Name })
+
+	case []interface{}:
+		for _, item := range v {
+			value, err := newValue(ctx, item)
+			if err != nil {
+				return nil, errs.Wrap(err, "Could not parse '%s' function's argument list item: %v", name, item)
+			}
+			args = append(args, value)
+		}
+
+	default:
+		return nil, errs.New("Function '%s' expected to be object or list", name)
+	}
+
+	return &Ap{Name: name, Arguments: args}, nil
+}
+
+func newAssignments(ctx context, m map[string]interface{}) (*[]*Var, error) {
+	ctx.push(ctxAssignments)
+	defer ctx.pop()
+
+	assignments := []*Var{}
+	for key, valueInterface := range m {
+		value, err := newValue(ctx, valueInterface)
+		if err != nil {
+			return nil, errs.Wrap(err, "Could not parse '%s' key's value: %v", key, valueInterface)
+		}
+		assignments = append(assignments, &Var{Name: key, Value: value})
+	}
+	sort.SliceStable(assignments, func(i, j int) bool { return assignments[i].Name < assignments[j].Name })
+	return &assignments, nil
+}
+
+func newIn(ctx context, inValue interface{}) (*In, error) {
+	ctx.push(ctxIn)
+	defer ctx.pop()
+
+	in := &In{}
+
+	switch v := inValue.(type) {
+	case map[string]interface{}:
+		f, err := newAp(ctx, v)
+		if err != nil {
+			return nil, errs.Wrap(err, "'in' object is not a function call")
+		}
+		in.FuncCall = f
+
+	case string:
+		in.Name = p.StrP(strings.TrimPrefix(v, "$"))
+
+	default:
+		return nil, errs.New("'in' value expected to be a function call or string")
+	}
+
+	return in, nil
 }
 
 // validateRequirements ensures that the requirements in the BuildExpression contain
 // both the name and namespace fields. These fileds are used for requirement operations.
-func validateRequirements(requirements []interface{}) error {
-	for _, requirement := range requirements {
-		r, ok := requirement.(map[string]interface{})
-		if !ok {
-			return errs.New("Requirement in BuildExpression is malformed")
+func (e *BuildExpression) validateRequirements() error {
+	requirements := e.getRequirementsNode()
+	for _, r := range requirements {
+		if r.Object == nil {
+			continue
 		}
 
-		_, ok = r[RequirementNameKey]
-		if !ok {
-			return errs.New("Requirement in BuildExpression is missing name field: %#v", r)
-		}
-		_, ok = r[RequirementNamespaceKey]
-		if !ok {
-			return errs.New("Requirement in BuildExpression is missing namespace field: %#v", r)
+		// The requirement object needs to have a name and value field.
+		// The value can be a string (in the case of name or namespace)
+		// or a list (in the case of version requirements).
+		for _, o := range *r.Object {
+			if o.Name == "" {
+				return errs.New("Requirement object missing name field")
+			}
+
+			if o.Value == nil {
+				return errs.New("Requirement object missing value field")
+			}
+
+			if o.Value.Str == nil && o.Value.List == nil {
+				return errs.New("Requirement object value field is not a string")
+			}
 		}
 	}
 	return nil
 }
 
 // Requirements returns the requirements in the BuildExpression.
-func (bx BuildExpression) Requirements() []Requirement {
-	return bx.requirements
-}
-
-// Update updates the BuildExpression's requirements based on the operation and requirement.
-func (bx *BuildExpression) Update(operation Operation, requirement Requirement, timestamp strfmt.DateTime) error {
-	var err error
-	switch operation {
-	case OperationAdded:
-		err = bx.addRequirement(requirement)
-	case OperationRemoved:
-		err = bx.removeRequirement(requirement)
-	case OperationUpdated:
-		err = bx.updateRequirement(requirement)
-	default:
-		return errs.New("Unsupported operation")
-	}
-	if err != nil {
-		return errs.Wrap(err, "Could not update BuildExpression's requirements")
-	}
-
-	formatted, err := time.Parse(time.RFC3339, timestamp.String())
-	if err != nil {
-		return errs.Wrap(err, "Could not parse latest timestamp")
-	}
-
-	(*bx.solveNode)[AtTimeKey] = formatted
-
-	return nil
-}
-
-// addRequirement adds a requirement to the BuildExpression.
-func (bx *BuildExpression) addRequirement(requirement Requirement) error {
-	bx.requirements = append(bx.requirements, requirement)
-
-	(*bx.solveNode)[RequirementsKey] = bx.requirements
-
-	return nil
-}
-
-// removeRequirement removes a requirement from the BuildExpression.
-func (bx *BuildExpression) removeRequirement(requirement Requirement) error {
-	for i, req := range bx.requirements {
-		if req.Name == requirement.Name && req.Namespace == requirement.Namespace {
-			bx.requirements = append(bx.requirements[:i], bx.requirements[i+1:]...)
-			(*bx.solveNode)[RequirementsKey] = bx.requirements
-			return nil
-		}
-	}
-
-	return errs.New("Could not find requirement")
-}
-
-// updateRequirement updates an existing requirement in the BuildExpression.
-func (bx BuildExpression) updateRequirement(requirement Requirement) error {
-	for i, req := range bx.requirements {
-		if req.Name == requirement.Name && req.Namespace == requirement.Namespace {
-			bx.requirements[i] = requirement
-			(*bx.solveNode)[RequirementsKey] = bx.requirements
-			return nil
-		}
-	}
-
-	return errs.New("Could not find requirement")
-}
-
-func (bx BuildExpression) MarshalJSON() ([]byte, error) {
-	return json.Marshal(bx.expression)
-}
-
-// getRequirements returns the list of requirements from the solve node of the build expression.
 // It returns an error if the requirements are not found or if they are malformed.
 // It expects the JSON representation of the solve node to be formatted as follows:
 //
@@ -418,37 +426,72 @@ func (bx BuildExpression) MarshalJSON() ([]byte, error) {
 //	    }
 //	  ]
 //	}
-func getRequirements(solveNode map[string]interface{}) ([]Requirement, error) {
-	for k, v := range solveNode {
-		if k != RequirementsKey {
+func (e *BuildExpression) Requirements() []Requirement {
+	requirementsNode := e.getRequirementsNode()
+
+	var requirements []Requirement
+	for _, r := range requirementsNode {
+		if r.Object == nil {
 			continue
 		}
 
-		node, ok := v.([]interface{})
-		if !ok {
-			return nil, errs.New("Requirements in BuildExpression are malformed")
-		}
+		var req Requirement
+		for _, o := range *r.Object {
+			if o.Name == RequirementNameKey {
+				req.Name = *o.Value.Str
+			}
 
-		requirementsData, err := json.Marshal(node)
-		if err != nil {
-			return nil, errs.Wrap(err, "Could not marshal JSON")
-		}
+			if o.Name == RequirementNamespaceKey {
+				req.Namespace = *o.Value.Str
+			}
 
-		var requirements []Requirement
-		err = json.Unmarshal(requirementsData, &requirements)
-		if err != nil {
-			return nil, errs.Wrap(err, "Could not unmarshal JSON")
+			if o.Name == RequirementVersionRequirementsKey {
+				req.VersionRequirement = getVersionRequirements(o.Value.List)
+			}
 		}
-
-		err = validateRequirements(node)
-		if err != nil {
-			return nil, errs.Wrap(err, "Requirements in BuildExpression are invalid")
-		}
-
-		return requirements, nil
+		requirements = append(requirements, req)
 	}
 
-	return nil, errs.New("Could not find requirements node")
+	return requirements
+}
+
+func (e *BuildExpression) getRequirementsNode() []*Value {
+	solveAp := e.getSolveNode()
+
+	var reqs []*Value
+	for _, arg := range solveAp.Arguments {
+		if arg.Assignment == nil {
+			continue
+		}
+
+		if arg.Assignment.Name == RequirementsKey && arg.Assignment.Value != nil {
+			reqs = *arg.Assignment.Value.List
+		}
+	}
+
+	return reqs
+}
+
+func getVersionRequirements(v *[]*Value) []VersionRequirement {
+	var reqs []VersionRequirement
+	for _, r := range *v {
+		if r.Object == nil {
+			continue
+		}
+
+		versionReq := make(VersionRequirement)
+		for _, o := range *r.Object {
+			if o.Name == RequirementComparatorKey {
+				versionReq[RequirementComparatorKey] = *o.Value.Str
+			}
+
+			if o.Name == RequirementVersionKey {
+				versionReq[RequirementVersionKey] = *o.Value.Str
+			}
+		}
+		reqs = append(reqs, versionReq)
+	}
+	return reqs
 }
 
 // getSolveNode returns the solve node from the build expression.
@@ -464,49 +507,224 @@ func getRequirements(solveNode map[string]interface{}) ([]Requirement, error) {
 //	    }
 //	  }
 //	}
-func getSolveNode(expression map[string]interface{}) (map[string]interface{}, error) {
-	solveNode, err := getFuncNode(expression, SolveFuncName)
-	if err == nil {
-		return solveNode, nil
-	}
-	if !errors.Is(err, funcNodeNotFoundError) {
-		return nil, errs.Wrap(err, "Could not get solve node")
-	}
-	logging.Debug("Could not get solve node, trying solve legacy node")
-
-	return getFuncNode(expression, SolveLegacyFuncName)
-}
-
-// getFuncNode returns the node of the given function name from the build expression.
-// It returns an error if the function node is not found.
-// Currently, this function just recurses the build expression until it finds the function node
-// of the correct map[string]interface{} type.
-// It expects the JSON representation of the build expression to be formatted as follows:
-//
-//	{
-//	  "let": {
-//	    "runtime": {
-//	      "func_name": {
-//	      }
-//	    }
-//	  }
-//	}
-func getFuncNode(expression map[string]interface{}, funcName string) (map[string]interface{}, error) {
-	for k, v := range expression {
-		node, ok := v.(map[string]interface{})
-		if !ok {
+func (e *BuildExpression) getSolveNode() *Ap {
+	for _, a := range e.Let.Assignments {
+		if a.Value.Ap == nil {
 			continue
 		}
 
-		if k == funcName {
-			return node, nil
-		}
-
-		// We recurse the build expression until we find the function node
-		if childNode, err := getFuncNode(node, funcName); err == nil {
-			return childNode, nil
+		if a.Value.Ap.Name == SolveFuncName || a.Value.Ap.Name == SolveLegacyFuncName {
+			return a.Value.Ap
 		}
 	}
 
-	return nil, funcNodeNotFoundError
+	return nil
+}
+
+// Update updates the BuildExpression's requirements based on the operation and requirement.
+func (e *BuildExpression) Update(operation Operation, requirement Requirement, timestamp strfmt.DateTime) error {
+	var err error
+	switch operation {
+	case OperationAdded:
+		err = e.addRequirement(requirement)
+	case OperationRemoved:
+		err = e.removeRequirement(requirement)
+	case OperationUpdated:
+		err = e.updateRequirement(requirement)
+	default:
+		return errs.New("Unsupported operation")
+	}
+	if err != nil {
+		return errs.Wrap(err, "Could not update BuildExpression's requirements")
+	}
+
+	err = e.updateTimestamp(timestamp)
+	if err != nil {
+		return errs.Wrap(err, "Could not update BuildExpression's timestamp")
+	}
+
+	return nil
+}
+
+func (e *BuildExpression) addRequirement(requirement Requirement) error {
+	obj := []*Var{
+		{Name: RequirementNameKey, Value: &Value{Str: p.StrP(requirement.Name)}},
+		{Name: RequirementNamespaceKey, Value: &Value{Str: p.StrP(requirement.Namespace)}},
+	}
+
+	if requirement.VersionRequirement != nil {
+		for _, r := range requirement.VersionRequirement {
+			obj = append(obj, &Var{Name: RequirementVersionRequirementsKey, Value: &Value{List: &[]*Value{
+				{Object: &[]*Var{
+					{Name: RequirementComparatorKey, Value: &Value{Str: p.StrP(r[RequirementComparatorKey])}},
+					{Name: RequirementVersionKey, Value: &Value{Str: p.StrP(r[RequirementVersionKey])}},
+				}}},
+			}})
+		}
+	}
+
+	requirementsNode := append(e.getRequirementsNode(), &Value{Object: &obj})
+
+	for _, arg := range e.getSolveNode().Arguments {
+		if arg.Assignment == nil {
+			continue
+		}
+
+		if arg.Assignment.Name == RequirementsKey {
+			arg.Assignment.Value.List = &requirementsNode
+		}
+	}
+
+	return nil
+}
+
+func (e *BuildExpression) removeRequirement(requirement Requirement) error {
+	requirementsNode := e.getRequirementsNode()
+
+	for i, r := range requirementsNode {
+		if r.Object == nil {
+			continue
+		}
+
+		for _, o := range *r.Object {
+			if o.Name == RequirementNameKey && *o.Value.Str == requirement.Name {
+				requirementsNode = append(requirementsNode[:i], requirementsNode[i+1:]...)
+			}
+		}
+	}
+
+	for _, arg := range e.getSolveNode().Arguments {
+		if arg.Assignment == nil {
+			continue
+		}
+
+		if arg.Assignment.Name == RequirementsKey {
+			arg.Assignment.Value.List = &requirementsNode
+		}
+	}
+
+	return nil
+}
+
+func (e *BuildExpression) updateRequirement(requirement Requirement) error {
+	requirementsNode := e.getRequirementsNode()
+
+	for _, r := range requirementsNode {
+		if r.Object == nil {
+			continue
+		}
+
+		for _, o := range *r.Object {
+			if o.Name == RequirementNameKey && *o.Value.Str == requirement.Name {
+				if requirement.VersionRequirement != nil {
+					for _, v := range *r.Object {
+						if v.Name == "version_requirements" {
+							v.Value.List = &[]*Value{
+								{Object: &[]*Var{
+									{Name: "comparator", Value: &Value{Str: p.StrP(requirement.VersionRequirement[0]["comparator"])}},
+									{Name: "version", Value: &Value{Str: p.StrP(requirement.VersionRequirement[0]["version"])}},
+								}},
+							}
+						}
+					}
+				}
+			}
+		}
+	}
+
+	for _, arg := range e.getSolveNode().Arguments {
+		if arg.Assignment == nil {
+			continue
+		}
+
+		if arg.Assignment.Name == RequirementsKey {
+			arg.Assignment.Value.List = &requirementsNode
+		}
+	}
+
+	return nil
+}
+
+func (e *BuildExpression) updateTimestamp(timestamp strfmt.DateTime) error {
+	formatted, err := time.Parse(time.RFC3339, timestamp.String())
+	if err != nil {
+		return errs.Wrap(err, "Could not parse latest timestamp")
+	}
+
+	for _, arg := range e.getSolveNode().Arguments {
+		if arg.Assignment == nil {
+			continue
+		}
+
+		if arg.Assignment.Name == "at_time" {
+			arg.Assignment.Value.Str = p.StrP(formatted.Format(time.RFC3339))
+		}
+	}
+
+	return nil
+}
+
+func (e *BuildExpression) MarshalJSON() ([]byte, error) {
+	m := make(map[string]interface{})
+	let := make(map[string]interface{})
+	for _, assignment := range e.Let.Assignments {
+		let[assignment.Name] = assignment.Value
+	}
+	m["let"] = let
+	return json.Marshal(m)
+}
+
+func (a *Var) MarshalJSON() ([]byte, error) {
+	m := make(map[string]interface{})
+	m[a.Name] = a.Value
+	return json.Marshal(m)
+}
+
+func (v *Value) MarshalJSON() ([]byte, error) {
+	switch {
+	case v.Ap != nil:
+		return json.Marshal(v.Ap)
+	case v.List != nil:
+		return json.Marshal(v.List)
+	case v.Str != nil:
+		return json.Marshal(strings.Trim(*v.Str, `"`))
+	case v.Null != nil:
+		return json.Marshal(nil)
+	case v.Assignment != nil:
+		return json.Marshal(v.Assignment)
+	case v.Object != nil:
+		m := make(map[string]interface{})
+		for _, assignment := range *v.Object {
+			m[assignment.Name] = assignment.Value
+		}
+		return json.Marshal(m)
+	case v.Ident != nil:
+		return json.Marshal(v.Ident)
+	}
+	return json.Marshal([]*Value{}) // participle does not create v.List if it's empty
+}
+
+func (f *Ap) MarshalJSON() ([]byte, error) {
+	m := make(map[string]interface{})
+	args := make(map[string]interface{})
+	for _, argument := range f.Arguments {
+		switch {
+		case argument.Assignment != nil:
+			args[argument.Assignment.Name] = argument.Assignment.Value
+		default:
+			return nil, fmt.Errorf("Cannot marshal %v (arg %v)", f, argument)
+		}
+	}
+	m[f.Name] = args
+	return json.Marshal(m)
+}
+
+func (i *In) MarshalJSON() ([]byte, error) {
+	switch {
+	case i.FuncCall != nil:
+		return json.Marshal(i.FuncCall)
+	case i.Name != nil:
+		return json.Marshal("$" + *i.Name)
+	}
+	return nil, fmt.Errorf("Cannot marshal %v", i)
 }
